Only create cups folder when it doesn't exist, so Configuration changes persist

+++ rootfs/run.sh
@@ -7,10 +7,12 @@ until [ -e /var/run/avahi-daemon/socket ]; do
 done
 
 bashio::log.info "Preparing directories"
-cp -v -R /etc/cups /data
-rm -v -fR /etc/cups
-
-ln -v -s /data/cups /etc/cups
+# Create if not exists
+if [ ! -d "/data/cups" ]; then
+  cp -v -R /etc/cups /data
+  rm -v -fR /etc/cups
+  ln -v -s /data/cups /etc/cups
+fi
 
 bashio::log.info "Starting CUPS server as CMD from S6"
